feat: move utils from matcher to recorder module - move decoding functions - modify approach when ending recording   - decode response right away
